move bubbletea output to separate util
